Time Series Analysis of Natural Gas Consumption
The aim of this study is to understand the behavior of the Natural Gas Consumption values by considering the past trends effects during given time frame. This study provides an in-depth look at time series data on natural gas use, improving our understanding of consumption patterns and providing valuable information on long-term trends and dynamics in the use of natural gas resources.